requirements.txt
psycopg2==2.9.3
psycopg2-binary==2.9.3
python-dotenv==1.0.0